Stop using plain np.float and np.int

+++ tests/datasets/tests/test_categorical.py
@@ -9,7 +9,7 @@ class TestRandomIntegerGenerator:
     def test(self):
         output = categorical.RandomIntegerGenerator.generate(10)
         assert len(output) == 10
-        assert output.dtype == np.int
+        assert output.dtype == int
         assert len(pd.unique(output)) < 6
         assert np.isnan(output).sum() == 0
 
@@ -19,7 +19,7 @@ class TestRandomIntegerNaNsGenerator:
     def test(self):
         output = categorical.RandomIntegerNaNsGenerator.generate(10)
         assert len(output) == 10
-        assert output.dtype == np.float
+        assert output.dtype == float
         assert len(pd.unique(output)) < 7
         assert np.isnan(output).sum() > 0
 
@@ -67,7 +67,7 @@ class TestSingleIntegerGenerator:
     def test(self):
         output = categorical.SingleIntegerGenerator.generate(10)
         assert len(output) == 10
-        assert output.dtype == np.int
+        assert output.dtype == int
         assert len(pd.unique(output)) == 1
         assert np.isnan(output).sum() == 0
 
@@ -77,7 +77,7 @@ class TestSingleIntegerNaNsGenerator:
     def test(self):
         output = categorical.SingleIntegerNaNsGenerator.generate(10)
         assert len(output) == 10
-        assert output.dtype == np.float
+        assert output.dtype == float
         assert len(pd.unique(output)) == 2
         assert np.isnan(output).sum() >= 1
 
@@ -107,7 +107,7 @@ class TestUniqueIntegerGenerator:
     def test(self):
         output = categorical.UniqueIntegerGenerator.generate(10)
         assert len(output) == 10
-        assert output.dtype == np.int
+        assert output.dtype == int
         assert len(pd.unique(output)) == 10
         assert np.isnan(output).sum() == 0
 
@@ -119,7 +119,7 @@ class TestUniqueIntegerNaNsGenerator:
         nulls = np.isnan(output).sum()
 
         assert len(output) == 10
-        assert output.dtype == np.float
+        assert output.dtype == float
         assert len(pd.unique(output)) == 10 - nulls + 1
         assert nulls > 0
